Bird Functional Traits Dataset (EltonTraits 1.0)

Dietary category percentages, foraging strategies, and functional traits for birds.

Usage

DBFunc_BirdFuncDat

Format

A data frame containing species dietary breakdowns (summing to 100

SpecID

Unique species identifier.

PassNonPass

Indicates whether the bird is a Passerine or Non-Passerine.

IOCOrder

IOC taxonomic order.

BLFamilyLatin

BirdLife International family Latin name.

BLFamilyEnglish

BirdLife International family English name.

BLFamSequID

BirdLife International family sequence ID.

Taxo

Taxonomic grouping code.

Scientific

Scientific name of the bird species.

English

Common English name of the bird species.

Diet.Inv

Percentage of invertebrates in the diet.

Diet.Vend

Percentage of vertebrate endotherms consumed.

Diet.Vect

Percentage of vertebrate ectotherms consumed.

Diet.Vfish

Percentage of fish consumed.

Diet.Vunk

Percentage of unknown/unclassified vertebrates in the diet.

Diet.Scav

Percentage of scavenging activity.

Diet.Fruit

Percentage of fruits consumed.

Diet.Nect

Percentage of nectar/pollen consumed.

Diet.Seed

Percentage of seeds/nuts consumed.

Diet.PlantO

Percentage of other plant tissues consumed.

Diet.5Cat

Simplified 5-category diet classification.

Diet.Source

Source of the dietary information.

Diet.Certainty

Certainty score for the diet data.

Diet.EnteredBy

Identifier of the person who entered the diet data.

ForStrat.watbelowsurf

Percentage of foraging time spent underwater.

ForStrat.wataroundsurf

Percentage of foraging time spent at the water surface.

ForStrat.ground

Percentage of foraging time spent on the ground.

ForStrat.understory

Percentage of foraging time spent in the understory.

ForStrat.midhigh

Percentage of foraging time spent in the mid-high canopy.

ForStrat.canopy

Percentage of foraging time spent in the upper canopy.

ForStrat.aerial

Percentage of foraging time spent in aerial foraging.

PelagicSpecialist

Indicator of pelagic specialization.

ForStrat.Source

Source of the foraging strategy data.

ForStrat.SpecLevel

Specificity level of the foraging strategy record.

ForStrat.EnteredBy

Identifier of the person who entered the foraging data.

Nocturnal

Indicator of nocturnal activity.

BodyMass.Value

Body mass value in grams.

BodyMass.Source

Source of the body mass information.

BodyMass.SpecLevel

Specificity level of the body mass record.

BodyMass.Comment

Notes on body mass data.

Record.Comment

General comments about the record.

Source

EltonTraits 1.0 database.